CheckpointEntryData moved into separate files.
