+++ src/checkpointing/CMakeLists.txt
@@ -9,7 +9,9 @@ set (PVLibSrcCpp ${PVLibSrcCpp}
 
 set (PVLibSrcHpp ${PVLibSrcHpp}
    ${SUBDIR}/CheckpointEntry.hpp
+   ${SUBDIR}/CheckpointEntryData.hpp
    ${SUBDIR}/CheckpointEntryDataStore.hpp
+   ${SUBDIR}/CheckpointEntryPvp.hpp
    ${SUBDIR}/CheckpointEntryRandState.hpp
    ${SUBDIR}/CheckpointEntryWeightPvp.hpp
    ${SUBDIR}/CheckpointableFileStream.hpp
